refactor(pbi-08-ai): AIClient委譲化, RemoteAIService slot loop, in-flight dedup
