+++ src/background/ai/aiServiceFactory.ts
@@ -23,6 +23,10 @@ export interface CreateAIServiceOptions {
 /**
  * Create the default AIService composition root:
  * FallbackAIService(local=LocalAIService, remote=RemoteAIService).
+ *
+ * The AIClient is retained only for backward compatibility with direct
+ * callers; the RemoteAIService it creates is reused as the remote service
+ * so there is exactly one slot loop instance.
  */
 export function createAIService(options: CreateAIServiceOptions = {}): AIService {
   const aiClient = options.aiClient || new AIClient();
@@ -31,7 +35,7 @@ export function createAIService(options: CreateAIServiceOptions = {}): AIService
   const localAiService = new LocalAIService({ localAiClient: builtInAiClient });
   const aiService = new FallbackAIService({
     local: localAiService,
-    remote: new RemoteAIService({ aiClient }),
+    remote: aiClient.remoteAiService,
   });
 
   return aiService;
